dashboard/requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-admin-bootstrapped==2.5.4
django-appconf==1.0.1
django-bootstrap-form==3.2
django-extensions==1.5.5
django-localflavor==1.1
django-model-utils==2.3.1
django-notifications-hq==0.7.1
django-user-accounts==1.2.0
djangorestframework==3.1.3
flup6==1.1.1
gunicorn==19.1.1
ipython==3.2.1
itsdangerous==0.24
Jinja2==2.7.3
jsonfield==1.0.3
MarkupSafe==0.23
patch==1.14.2
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
static3==0.6.1
Werkzeug==0.10.4
whitenoise==2.0.3
